+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * nostr: use `OnceCell` implementation from `std` lib instead of `once_cell` ([Yuki Kishimoto])
 * nostr: remove redundant public key from repost events ([Yuki Kishimoto])
 * nostr: change `impl Ord for Event` behaviour (descending order instead of ascending) ([Yuki Kishimoto])
+* nostr: change `TagStandard::Relays` variant value from `Vec<UncheckedUrl>` to `Vec<Url>` ([Yuki Kishimoto])
 * relay-builder: refactor `Session::check_rate_limit` method ([Yuki Kishimoto])
 * relay-builder: return error if event was deleted ([Yuki Kishimoto])
 * pool: changes in `RelayPool::remove_relay` behavior ([Yuki Kishimoto])